Send Minikit and True Jedi checks as soon as they are achieved
The original checks, by reading the save file, are still present, to
allow for playing after losing connection to the Archipelago server, and
then reconnecting and sending checks.

The True Jedi check is slightly incorrect in that it will only send once
the player has collected a stud after reaching the required amount,
rather than sending as soon as the required amount has been reached.