Fix GOG compatibility and in-game message display not disabling
The known GOG offset of 32 bytes only actually applies to memory
addresses after a certain point, so the client has been updated to only
offset addresses greater than that point. The exact point is unknown,
but all memory addresses used by the client have been checked for
whether they need the offset or do not.

Because the bytes used to detect the game version were in a part of
memory with a different offset to the usual 32 bytes, the game version
detection was not working. The game version detection has been changed
to use a different pattern whose found memory address differs between
the Steam and GOG versions.

--

If the in-game messages cannot find the bytes pattern and determine the
localization language, then in-game messages should have been disabled,
but they were not being disabled.