Fix in-level True Jedi sending early/late
The address of a byte that is set to 1 once True Jedi has been completed
has been found, so the client now checks that byte.

In the previous version, it seems that losing studs could result in a
race condition where the P1 and P2 studs could be retrieved from memory
and before retrieving the True Jedi progress studs from memory, the
player would lose studs, resulting in the sum of P1 and P2 studs
exceeding the True Jedi studs, causing the check to send early.