Further improvements to client/game/server connection flow
Connecting to a server before the game is started is now allowed.

Fix successfully connecting on a save file, then loading a new save, and
then loading a save for a different multiworld not aborting the
connection.

Add a client and in-game message for when the client+game are fully
connected and communicating.