Re-write slot name and seed hash if the save data is rolled back
There is an edge-case where the player can connect to a slot on a file
that has already been saved, writing the slot name and seed hash into
the save data, but before the save data is persisted, the player could
return to the main menu and then load that same save file again. This
results in the slot name and seed hash being absent from save data, but
the client could still be connected to the server at that point.

An extra check is now run when the player goes from being not being
loaded into a save slot to being loaded into a save slot, that the slot
name/seed hash is present. If not, the slot name/seed hash is rewritten
into the save data.