Fix giving all True Jedi Gold Bricks for disabled locations for free
The client was seeing that a True Jedi location was not enabled, and
then marking the True Jedi as complete in datastorage, which the client
would read and set the complete flags for because that is needed to
better support same-slot co-op.

The client should not have been marking the True Jedi as complete in
datastorage to begin with.